backend/requirements.txt
Django>=5.2
djangorestframework>=3.16.0
django-cors-headers>=4.7.0
djangorestframework-simplejwt>=5.5.0
django-filter>=25.1
Pillow>=11.1.0
python-dotenv>=1.0.0
